# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Church Youth Group Coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/invite:
    post:
      operationId: inviteUser
      tags: [users]
      summary: Invite a new member (Presidency or Leader)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteUserInput"
      responses:
        "201":
          description: Invited user with access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user (role, status, group, access code)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}/reset-code:
    post:
      operationId: resetAccessCode
      tags: [users]
      summary: Reset a user's access code
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetCodeInput"
      responses:
        "200":
          description: Updated user with new code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ACTIVITIES
...
components:
  schemas:
...
    User:
      type: object
      required: [id, firstName, lastName, role, groupId, groupName, status, accessCode]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [member, presidency, leader]
        groupId:
          type: integer
        groupName:
          type: string
        status:
          type: string
          enum: [active, inactive, archived]
        accessCode:
          type: string
        createdAt:
          type: string
    InviteUserInput:
      type: object
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [member, presidency, leader]
        customAccessCode:
          type: ["string", "null"]
    UserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [member, presidency, leader]
        status:
          type: string
          enum: [active, inactive, archived]
        groupId:
          type: ["integer", "null"]
    ResetCodeInput:
      type: object
      properties:
        customCode:
          type: ["string", "null"]

    # ACTIVITIES
...