# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church Youth Group Coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  role: UserRole;
  groupId: number;
  groupName: string;
  status: UserStatus;
  accessCode: string;
  createdAt?: string;
}