# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Church Youth Group Coordination API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';
import type { UserUpdateStatus } from './userUpdateStatus';

export interface UserUpdate {
  role?: UserUpdateRole;
  status?: UserUpdateStatus;
  /** @nullable */
  groupId?: number | null;
}